Introduce StopOnMaxLength. (#1882)
`Model.sample_decode` by default continues decoding until it reaches the end of
the given prefix tensor’s `seq_len`. However, decoding can be stopped earlier
using a **stop_decoding_condition** arg.

For example, if a decoding evaler knows the **maximum generation length** for
an eval set, it may want to stop decoding once that limit is reached. To
support this, we introduce **`StopOnMaxLength`**.

For example, in a QA task, the question prefix length can vary significantly
across examples, but generating 30 tokens for the answer is usually sufficient.
In an ASR task, the audio prefix length can range widely from 100 to 1500, yet
a transcription of 150 tokens is enough.
Until the model is well-trained and learns to produce the EOS token
appropriately, the decoding evaler wastes time by generating up to the full
seq_len. This significantly speeds up decoding evalers.

In addition, introduce **CompositeDecodingCondition** to combine
**StopOnMaxLength** with **StopOnSubsequence**.

Also, removed `out_of_prompt` from **StopDecodingCondition**. This should be
handled by the **caller**, as it is not information that
**StopDecodingCondition** needs to know.