Change the return type hint of `get_summaries()` to `dict[str, MetricSummary]`. (#1891)
Instead of `dict[str, WeightedScalar]`.
Downstream code also uses **`MaxSummary`** and **`MinSummary`** for metrics,
which are not `WeightedScalar`.
Therefore, we define a unified **`MetricSummary`** type that encompasses all
metric variants and use `MetricSummary` as the type hint for all metric-related
interfaces.

Additionally, we add **`SumSummary`** to support downstream use cases.