Relax dtype constraint for `compute_conv(_transpose)_paddings`. (#2020)
This function was originally designed to take boolean paddings as input, but
downstream we also want to use `segment_ids` as input. Since the implementation
only samples anchor points and does not depend on the input dtype, it is safe
to relax the dtype constraint.

In addition, `ConvXXXWithPadding` provides a `conv_paddings()` method so that
downstream code can conv-op `paddings` and `segment_ids` easily.